+++ src/components/SocialMediaLinks/socialMediaLinksChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2024-11-07",
+    version: "3.4.2",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: ["Added support for the `socialX` icon."],
+  },
+  {
     date: "2024-07-25",
     version: "3.2.0",
     type: "Update",