+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ========
 
 ## Prerelease
+ - Adds the `useFormatNumber` hook.
 
 ## 3.5.1 (December 19, 2024)