+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates the `Heading` component to add explicit values for the mobile `line-height` styles.
+
 ## 3.5.2 (January 16, 2025)
 
 ### Adds