+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates the `Pagination` component to handle 4 digit page counts.
+- Updates the `Heading` component to add `line-height` styles for mobile.
 - Updates the `Breadcrumbs` component to make link text screen readable on mobile breakpoints.
 
 ## 3.5.2 (January 16, 2025)